categories table improvement

+++ database/migrations/2026_07_10_142227_create_categories_table.php
@@ -14,7 +14,7 @@ return new class extends Migration {
             $table->id();
             $table->string("name");
             $table->string("slug")->unique();
-            $table->integer("products_count")->default(0);
+            $table->unsignedBigInteger("products_count")->default(0);
             $table->timestamps();
             $table->softDeletes();
         });
